pyproject.toml
[project]
name = "nucleo-logico-evolutivo"
version = "0.1.0"
description = "Nucleo Logico Evolutivo: Interfaz Adaptativa entre LLM, Lean 4 y Grafo Categorico de Skills"
authors = [
    { name = "Leonardo Jimenez Martinez", email = "leonardo@unam.mx" }
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
keywords = ["theorem-proving", "lean4", "reinforcement-learning", "category-theory"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Mathematics",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Core
    "pydantic>=2.0",
    "typing-extensions>=4.0",

    # ML/RL
    "torch>=2.0",
    "torch-geometric>=2.3",
    "gymnasium>=0.29",
    "stable-baselines3>=2.0",

    # LLM Integration
    "anthropic>=0.18",

    # Graph Processing
    "networkx>=3.0",
    "numpy>=1.24",

    # Utilities
    "rich>=13.0",
    "pyyaml>=6.0",
    "aiofiles>=23.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1",
    "mypy>=1.0",
    "pre-commit>=3.0",
]
docs = [
    "mkdocs>=1.5",
    "mkdocs-material>=9.0",
    "mkdocstrings[python]>=0.24",
]

[project.scripts]
nucleo = "nucleo.cli:main"

[project.urls]
Homepage = "https://github.com/ai-enhanced-engineer/metamath-prover"
Documentation = "https://metamathematics.ai/docs"
Repository = "https://github.com/ai-enhanced-engineer/metamath-prover"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["nucleo"]

[tool.black]
line-length = 88
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 88
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=nucleo --cov-report=term-missing"
